refactor: move introduction section to its own component
